+++ src/main/java/com/ajouevent/admin/config/WebConfig.java
@@ -7,6 +7,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.session.web.http.HttpSessionIdResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.session.web.http.HeaderHttpSessionIdResolver;
 
@@ -32,6 +33,22 @@ public class WebConfig implements WebMvcConfigurer {
                 .exposedHeaders("X-Session-Id")
                 .allowCredentials(true); // 쿠키 포함 허용!
     }
+
+    //react 정적으로 때렸을 때, dom 오류 방지
+    @Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        // [a-zA-Z0-9-_]+ 1단계 경로 매핑
+        registry.addViewController("/{spring:[a-zA-Z0-9-_]+}")
+                .setViewName("forward:/index.html");
+        // 2단계 경로 매핑
+        registry.addViewController("/{spring:[a-zA-Z0-9-_]+}/{spring2:[a-zA-Z0-9-_]+}")
+                .setViewName("forward:/index.html");
+        // 3단계 경로 매핑
+        registry.addViewController("/{spring:[a-zA-Z0-9-_]+}/{spring2:[a-zA-Z0-9-_]+}/{spring3:[a-zA-Z0-9-_]+}")
+                .setViewName("forward:/index.html");
+    }
+
+
     @Bean
     public FilterRegistrationBean<AdminAuthCheckFilter> adminAuthFilterRegistration() {
         FilterRegistrationBean<AdminAuthCheckFilter> registrationBean = new FilterRegistrationBean<>();
